Pullng the C-API functions for graphics transforms into the RenderManagerC files. Simplified structure of remaining files, deleting ones no longer needed.
