+++ osvr/RenderKit/RenderManagerC.h
@@ -67,7 +67,7 @@ typedef struct OSVR_RenderParams {
 
 //=========================================================================
 /// Description needed to construct an off-axes projection matrix
-typedef struct OSVR_ProjectionMatrix {
+typedef struct {
     double left;
     double right;
     double top;
@@ -184,6 +184,99 @@ OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode
 osvrRenderManagerGetNumRenderInfoInCollection(
     OSVR_RenderInfoCollection renderInfoCollection,
     OSVR_RenderInfoCount* countOut);
+
+
+//=========================================================================
+// Routines to turn the OSVR_PoseState into ModelView matrices for various
+// rendering engines, including OpenGL, Direct3D, Unreal, and Unity.
+// Done in such a way that we don't require the inclusion
+// of the native API header files.
+
+/// Produce an OpenGL ModelView matrix from an OSVR_PoseState.
+/// Assumes that the world is described in a right-handed fashion and
+/// that we're going to use a right-handed projection matrix.
+/// @brief Produce OpenGL ModelView transform from OSVR_PoseState
+/// @param state_in Input state from RenderManager.
+/// @param OpenGL_out Pointer to 16-element double array that has
+///        been allocated by the caller.
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_PoseState_to_OpenGL(
+  double* OpenGL_out, const OSVR_PoseState& state_in);
+
+/// Produce a D3D ModelView matrix from an OSVR_PoseState.
+/// Handles transitioning from the right-handed OSVR coordinate
+/// system to the left-handed projection matrix that is typical
+/// for D3D applications.
+/// @brief Produce D3D ModelView transform from OSVR_PoseState
+/// @param state_in Input state from RenderManager.
+/// @param OpenGL_out Pointer to 16-element double array that has
+///        been allocated by the caller.
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_PoseState_to_D3D(
+  float D3D_out[16], const OSVR_PoseState& state_in);
+
+/// Modify the OSVR_PoseState from OSVR to be appropriate for use
+/// in a Unity application.  OSVR's world is right handed, and Unity's
+/// is left handed.
+/// @brief Modify OSVR_PoseState for use by Unity.
+/// @param state_in Input state from RenderManager.
+/// @param state_out Ouput state for use by Unity
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_PoseState_to_Unity(
+  OSVR_PoseState& state_out, const OSVR_PoseState& state_in);
+
+//=========================================================================
+// Routines to turn the 4x4 projection matrices returned as part of the
+// RenderCallback class into Projection matrices for OpenGL and
+// Direct3D.  Done in such a way that we don't require the inclusion of the
+// native API header files (since most apps will not include all of the
+// libraries).
+
+/// Produce an OpenGL Projection matrix from an OSVR_ProjectionMatrix.
+/// Assumes that the world is described in a right-handed fashion and
+/// that we're going to use a right-handed projection matrix.
+/// @brief Produce OpenGL Projection matrix from OSVR projection info.
+/// @param projection_in Input projection description from RenderManager.
+/// @param OpenGL_out Pointer to 16-element double array that has
+///        been allocated by the caller.
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_Projection_to_OpenGL(
+  double* OpenGL_out, const OSVR_ProjectionMatrix& projection_in);
+
+/// Produce a D3D Projection matrix from an OSVR_ProjectionMatrix.
+/// Produces a left-handed projection matrix as is typical
+/// for D3D applications.
+/// @brief Produce D3D Projection transform from OSVR projection info
+/// @param projection_in Input projection description from RenderManager.
+/// @param D3D_out Pointer to 16-element float array that has
+///        been allocated by the caller.
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_Projection_to_D3D(
+  float D3D_out[16], const OSVR_ProjectionMatrix& projection_in);
+
+/// Produce an Unreal Projection matrix from an OSVR_ProjectionMatrix.
+/// Produces a left-handed projection matrix whose Z values are
+/// in the opposite order, with Z=0 at the far clipping plane and
+/// Z=1 at the near clipping plane.  If there is not a far clipping
+/// plane defined, then set it to be the same as the near
+/// clipping plane before calling this function.  If there is not a
+/// near clipping plane set, then set it to 1 before calling this
+/// function.
+/// To put the result into an Unreal FMatrix, do the following:
+///   float p[16];
+///   OSVR_Projection_to_D3D(p, projection_in);
+///   FPlane row1(p[0], p[1], p[2], p[3]);
+///   FPlane row2(p[4], p[5], p[6], p[7]);
+///   FPlane row3(p[8], p[9], p[10], p[11]);
+///   FPlane row4(p[12], p[13], p[14], p[15]);
+///   FMatrix ret = FMatrix(row1, row2, row3, row4);
+/// @brief Produce Unreal Projection transform from OSVR projection info
+/// @param projection_in Input projection description from RenderManager.
+/// @param Unreal_out Pointer to 16-element float array that has
+///        been allocated by the caller.
+/// @return True on success, false on failure (null pointer).
+OSVR_RENDERMANAGER_EXPORT OSVR_ReturnCode OSVR_Projection_to_Unreal(
+  float Unreal_out[16], const OSVR_ProjectionMatrix& projection_in);
 
 OSVR_EXTERN_C_END
